+++ src/pages/home/Skills.js
@@ -1,9 +1,11 @@
-import React, { useState } from 'react';
+import React, { useEffect, useState } from 'react';
 import { Link } from 'react-router-dom';
 import { useSelector } from 'react-redux';
 import clsx from 'clsx';
 
 import { Container, Checkbox, FormControlLabel, makeStyles, withStyles, Button } from '@material-ui/core';
+
+import { chunkanizeArray, getRandomId } from 'utils/helpers';
 
 const useStyles = makeStyles((theme) => ({
     button: {
@@ -16,6 +18,7 @@ const useStyles = makeStyles((theme) => ({
     },
     pageTitle: {
         textTransform: 'uppercase',
+        fontSize: '1.8rem',
         [theme.breakpoints.up('md')]: {
             fontSize: '2.5rem',
             width: '65%',
@@ -28,14 +31,65 @@ const useStyles = makeStyles((theme) => ({
     root: {
         minHeight: '100vh',
         color: theme.palette.primary.main,
-        padding: '4rem 0'
+        paddingTop: '4rem',
+        paddingBottom: '4rem'
     },
-    skillsGrid: {
-        display: 'grid',
-        gridTemplateColumns: 'repeat(3, 1fr)',
-        gap: '.2rem 8rem',
-        marginTop: '5rem'
-    },
+    skillsGrid: (props) => ({
+        display: 'flex',
+        marginTop: '5rem',
+        gap: '.2rem 2rem',
+        overflowX: 'scroll',
+        scrollbarWidth: 'none',
+        scrollBehavior: 'smooth',
+        scrollSnapType: 'x mandatory',
+        [theme.breakpoints.up('md')]: {
+            display: 'grid',
+            gridTemplateColumns: `repeat(${props.skillsGroups}, 1fr)`,
+            gap: '.2rem 1rem'
+        },
+        '&::-webkit-scrollbar': {
+            width: 0,
+            height: 0
+        },
+        '&>div': {
+            display: 'flex',
+            flexDirection: 'column',
+            width: 'calc(100vw - 24px)',
+            height: 'calc(42px * 6 + 10px + 2rem)',
+            flexShrink: 0,
+            scrollSnapAlign: 'start',
+            scrollSnapStop: 'always',
+            padding: '0 10vw',
+            [theme.breakpoints.up('md')]: {
+                width: 'auto',
+                padding: '0'
+            },
+            '& .dots': {
+                display: 'flex',
+                justifyContent: 'center',
+                alignItems: 'flex-end',
+                marginTop: '2rem',
+                height: '100%',
+                [theme.breakpoints.up('md')]: {
+                    display: 'none'
+                },
+                '&>div': {
+                    display: 'flex',
+                    justifyContent: 'center',
+                    '&>span': {
+                        width: '10px',
+                        height: '10px',
+                        borderRadius: '100%',
+                        border: `2px solid ${theme.palette.primary.main}`,
+                        margin: '0 1rem',
+                        '&.fill': {
+                            backgroundColor: theme.palette.primary.main
+                        }
+                    }
+                }
+            }
+        }
+    }),
     underlineLink: {
         borderBottom: `2px solid ${theme.palette.primary.main}`
     }
@@ -52,13 +106,18 @@ const CustomCheckbox = withStyles((theme) => ({
 }))((props) => <Checkbox color="default" {...props} />);
 
 const Skills = ({ classes, ...props }) => {
-    const internalClasses = useStyles();
-
     const [checkedSkills, setCheckedSkills] = useState({});
+    const [chunkanizedSkills, setChunkanizedSkills] = useState([]);
+
+    const internalClasses = useStyles({ skillsGroups: chunkanizedSkills?.length });
 
     const textProvider = useSelector(({ ui }) => ui.textContent?.homePage?.skills);
     const currentLanguage = useSelector(({ ui }) => ui.appSettings?.currentLanguage);
     const skills = useSelector(({ entities }) => entities.skills?.list);
+
+    useEffect(() => {
+        if (skills) setChunkanizedSkills(chunkanizeArray(skills));
+    }, []);
 
     const handleCheckSkill = (event) => {
         if (!event.target.checked) {
@@ -76,18 +135,29 @@ const Skills = ({ classes, ...props }) => {
             </h2>
 
             <div className={internalClasses.skillsGrid}>
-                {skills?.map((skill) => (
-                    <FormControlLabel
-                        key={skill?.id}
-                        control={
-                            <CustomCheckbox
-                                onChange={handleCheckSkill}
-                                checked={checkedSkills?.[skill?.id] || false}
-                                name={skill?.id}
+                {chunkanizedSkills?.map((chunk, index) => (
+                    <div key={getRandomId()}>
+                        {chunk?.map((skill) => (
+                            <FormControlLabel
+                                key={skill?.id}
+                                control={
+                                    <CustomCheckbox
+                                        onChange={handleCheckSkill}
+                                        checked={checkedSkills?.[skill?.id] || false}
+                                        name={skill?.id}
+                                    />
+                                }
+                                label={skill?.title || skill?.[currentLanguage]?.title}
                             />
-                        }
-                        label={skill?.title || skill?.[currentLanguage]?.title}
-                    />
+                        ))}
+                        <div className="dots">
+                            <div>
+                                {chunkanizedSkills?.map((chunk, index2) => (
+                                    <span key={getRandomId()} className={index === index2 ? 'fill' : ''}></span>
+                                ))}
+                            </div>
+                        </div>
+                    </div>
                 ))}
             </div>